*添加了UIScroll,可以创建能够随用户鼠标滚动的图像UI *现在对着有碰撞体积的UI滚动鼠标不会缩放地图 *添加了能量条UI,可以显示金沙能量的多少

+++ idea_UI_button_menu.cpp
@@ -17,7 +17,8 @@ idea_UI_button_menu::idea_UI_button_menu()
 	texture_pressed = RenderingSystem::getInstance()->getAnimation(AnimationType::ui_button_menu_pressed, 0);
 	rendering_unit->setTexture(texture_default);
 
-	setPosition(WINDOW_WIDTH - 9 * UI_SCALE - 16, 16, 9 * UI_SCALE, 9 * UI_SCALE);
+	setPosition(WINDOW_WIDTH - 9 * static_cast<int>(UI_SCALE) - 16, 16,
+		9 * static_cast<int>(UI_SCALE), 9 * static_cast<int>(UI_SCALE));
 }
 
 idea_UI_button_menu::~idea_UI_button_menu()
